Fix assertion when test fast/events/ios/rotation/basic-rotation.html is followed by fast/events/ios/rotation/do-not-shrink-to-fit-content-after-rotation.html https://bugs.webkit.org/show_bug.cgi?id=206383 <rdar://problem/58618513>
Reviewed by Tim Horton.

Rotation tests are asserting because _didEndRotation is being called multiple times.
The call is at the end of viewWillTransitionToSize, which is called with the same
size that the view already is. Therefore, if no size is changing, early out,
which keeps this double call from happening.

* WebKitTestRunner/ios/PlatformWebViewIOS.mm:
(-[PlatformWebViewController viewWillTransitionToSize:withTransitionCoordinator:]):


Identifier: 219510@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254765 268f45cc-cd09-0410-ab3c-d52691b4dbfc