[GTK][WPE] TestWebKitUserContentFilterStore is crashing https://bugs.webkit.org/show_bug.cgi?id=206527
Patch by Alejandro G. Castro <alex@igalia.com> on 2020-01-21
Reviewed by Adrian Perez de Castro.

We are asserting looking for a category of error when converting
in the toGError function, but depending on the function called the
category can be different.

* UIProcess/API/glib/WebKitUserContentFilterStore.cpp:
(toGError):
(webkitUserContentFilterStoreSaveBytes): Moved the ASSERT to the
function that uses this category, the other functions calling
toGError were already ASSERTing the correct category.

Identifier: 219589@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254852 268f45cc-cd09-0410-ab3c-d52691b4dbfc